chore(deps): bump the gradle-minor-patch group in /mobile with 2 updates

+++ mobile/shared/src/commonMain/kotlin/com/bissbilanz/analytics/Tdee.kt
@@ -48,7 +48,7 @@ private class DatedPoint(
     val value: Double,
 )
 
-private fun epochDay(date: String): Int = LocalDate.parse(date).toEpochDays()
+private fun epochDay(date: String): Int = LocalDate.parse(date).toEpochDays().toInt()
 
 /**
  * OLS slope of value on calendar day (kg/day). Regressing on the date rather
